+++ deal.II/contrib/mesh_conversion/MeshConversion.h
@@ -23,7 +23,7 @@ class MeshConversion
 
 	private:
 		void greeting ();
-		std::vector <double> get_global_node_numbers(const int face_cell_no, const int face_cell_face_no);
+		std::vector <int> get_global_node_numbers(const int face_cell_no, const int face_cell_face_no);
 		
 		const double 				tolerance;
 		const unsigned int 			dimension;
@@ -36,8 +36,8 @@ class MeshConversion
 		std::string 				output_file_name;
 		
 		std::vector< std::vector<double> > 	node_list; 	// Stored as [ global node-id (int), x-coord, y-coord, z-coord ]
-		std::vector< std::vector<double> > 	cell_list; 	// Stored as [ material-id (int), node1, node2, node3, node4, node5, node6, node7, node8 ]
-		std::vector< std::vector<double> > 	face_list; 	// Stored as [ sideset-id (int), node1, node2, node3, node4 ]
+		std::vector< std::vector<int> > 	cell_list; 	// Stored as [ material-id (int), node1, node2, node3, node4, node5, node6, node7, node8 ]
+		std::vector< std::vector<int> > 	face_list; 	// Stored as [ sideset-id (int), node1, node2, node3, node4 ]
 };
 
 #endif // MESH_CONVERSION_H
